refactor: moved overview panel logic to own file
