+++ site/src/components/QueueItem.svelte
@@ -1,20 +1,46 @@
+<script context="module" lang="ts">
+export interface QueueTitleInfo {
+    Title: string
+    Episodic: boolean
+    Episode: number
+    Season: number
+    Year: number
+    Resolution: string
+}
+
+export interface QueueOmdbInfo {
+    Title: string
+    plot: string
+    ReleaseYear: number
+    Runtime: string
+    Type: string
+    poster: string
+    ImdbId: string
+    Response: boolean
+    Error: string
+    Genre: string[]
+}
+
+export interface QueueTroubleInfo {
+    message: string
+}
+
+export interface QueueDetails extends QueueItem {
+    title_info: QueueTitleInfo
+    omdb_info:  QueueOmdbInfo
+    trouble: QueueTroubleInfo
+}
+</script>
+
 <script lang="ts">
 import { onMount } from "svelte";
 import { commander } from "../commander";
-import { SocketMessageType, SocketPacketType } from "../store";
+import { SocketMessageType } from "../store";
 import type { SocketData } from "../store";
 import type { QueueItem } from "./Queue.svelte";
 
-import importIcon from '../assets/import-stage.svg';
-import titleIcon from '../assets/title-stage.svg';
-import omdbIcon from '../assets/omdb-stage.svg';
-import ffmpegIcon from '../assets/ffmpeg-stage.svg';
-import dbIcon from '../assets/db-stage.svg';
 import troubleIcon from '../assets/warning.svg';
-
-import ellipsisHtml from '../assets/html/ellipsis.html';
-import workingHtml from '../assets/html/dual-ring.html';
-import errHtml from '../assets/err.svg';
+import OverviewPanel from "./panels/OverviewPanel.svelte";
 
 enum ComponentState {
     LOADING,
@@ -30,39 +56,7 @@ enum ComponentPage {
     DB,
 }
 
-interface QueueTitleInfo {
-    Title: string
-    Episodic: boolean
-    Episode: number
-    Season: number
-    Year: number
-    Resolution: string
-}
 
-interface QueueOmdbInfo {
-    Title: string
-    plot: string
-    ReleaseYear: number
-    Runtime: string
-    Type: string
-    poster: string
-    ImdbId: string
-    Response: boolean
-    Error: string
-    Genre: string[]
-}
-
-interface QueueTroubleInfo {
-    message: string
-}
-
-interface QueueDetails extends QueueItem {
-    title_info: QueueTitleInfo
-    omdb_info:  QueueOmdbInfo
-    trouble: QueueTroubleInfo
-}
-
-const els:HTMLElement[] = new Array(4);
 export let queueInfo:QueueItem
 
 let state = ComponentState.LOADING
@@ -79,7 +73,6 @@ const getDetails = function() {
             details = response.arguments.payload
             state = ComponentState.COMPLETE
 
-            requestAnimationFrame(updateEllipsis);
 
             return true
         }
@@ -108,21 +101,6 @@ const getStatusStr = function(status:number): string {
         case 2: return "Completed"
         case 3: return "Troubled"
     }
-}
-
-const updateEllipsis = function() {
-    const left = els[details.stage] as HTMLElement
-    const right = els[details.stage + 1] as HTMLElement
-    const spinner = els[5];
-
-    if (!left || !right || !spinner) {
-        return
-    }
-
-    const mid = ((left.offsetLeft + left.offsetWidth) + (right.offsetLeft))/2
-    spinner.setAttribute("style", "left: " + (mid - spinner.offsetWidth / 2) + "px;")
-
-    requestAnimationFrame(updateEllipsis)
 }
 
 $:stat = function() {
@@ -167,22 +145,7 @@ onMount(getDetails)
             <span class:active="{page == ComponentPage.DB}" on:click="{() => page = ComponentPage.DB}">DB</span>
         </div>
         <main>
-            <div class="stages">
-                <div bind:this={els[0]} class="stage import" class:hidden="{0 < details.stage}" class:active="{details.stage == 0}"><span class="caption">Import</span>{@html importIcon}</div>
-                <div bind:this={els[1]} class="stage title" class:hidden="{1 < details.stage}" class:active="{details.stage == 1}"><span class="caption">Title</span>{@html titleIcon}</div>
-                <div bind:this={els[2]} class="stage omdb" class:hidden="{details.stage == 0 || details.stage > 2}" class:active="{details.stage == 2}"><span class="caption">OMDB</span>{@html omdbIcon}</div>
-                <div bind:this={els[3]} class="stage ffmpeg" class:hidden="{details.stage > 1 && details.stage < 2}" class:active="{details.stage == 3}"><span class="caption">Ffmpeg</span>{@html ffmpegIcon}</div>
-                <div bind:this={els[4]} class="stage db" class:hidden="{details.stage < 3}" class:active="{details.stage == 4}"><span class="caption">DB</span>{@html dbIcon}</div>
-                <div bind:this={els[5]} class="loading">
-                    {#if details.trouble}
-                        {@html errHtml}
-                    {:else if details.status == 0}
-                        {@html ellipsisHtml}
-                    {:else if details.status == 1}
-                        {@html workingHtml}
-                    {/if}
-                </div>
-            </div>
+            <OverviewPanel details={details}/>
         </main>
     </div>
 {:else}
